Use native V8 printing facilities
V8 properly escapes log lines by making sure that they don't contain
'\n' or ',' (replaced by literal '\n' and '\x2C' respectively).  This
greatly simplifies the parsing of log files in the analyzer.

Also reformats the log.h and log.cc files to be more consistent with the
rest of the codebase.